restructured converters to their own folder
