+++ StudySmarterFlashcards/App.xaml
@@ -2,7 +2,6 @@
     x:Class="StudySmarterFlashcards.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:StudySmarterFlashcards"
     xmlns:utils="using:StudySmarterFlashcards.Utils">
   <Application.Resources>
     <ResourceDictionary>
