+++ StudySmarterFlashcards/Converters/BooleanToStarFillConverter.cs
@@ -1,17 +1,15 @@
 ﻿using System;
-using Windows.UI.Xaml;
 using Windows.UI.Xaml.Data;
 
-namespace StudySmarterFlashcards.Utils
+namespace StudySmarterFlashcards.Converters
 {
-  public class BooleanToVisibilityConverter : IValueConverter
+  public class BooleanToStarFillConverter : IValueConverter
   {
     public object Convert(object value, Type targetType, object parameter, string language) =>
         (bool)value ^ (parameter as string ?? string.Empty).Equals("Reverse") ?
-            Visibility.Visible : Visibility.Collapsed;
+            "\xE735" : "\xE734";
 
     public object ConvertBack(object value, Type targetType, object parameter, string language) =>
-        (Visibility)value == Visibility.Visible ^ (parameter as string ?? string.Empty).Equals("Reverse");
-
+        ((string)value).Equals("\xE735") ^ (parameter as string ?? string.Empty).Equals("Reverse");
   }
 }
