+++ StudySmarterFlashcards/Converters/BooleanToVisibilityConverter.cs
@@ -1,15 +1,17 @@
 ﻿using System;
+using Windows.UI.Xaml;
 using Windows.UI.Xaml.Data;
 
-namespace StudySmarterFlashcards.Utils
+namespace StudySmarterFlashcards.Converters
 {
-  public class BooleanToStarFillConverter : IValueConverter
+  public class BooleanToVisibilityConverter : IValueConverter
   {
     public object Convert(object value, Type targetType, object parameter, string language) =>
         (bool)value ^ (parameter as string ?? string.Empty).Equals("Reverse") ?
-            "\xE735" : "\xE734";
+            Visibility.Visible : Visibility.Collapsed;
 
     public object ConvertBack(object value, Type targetType, object parameter, string language) =>
-        ((string)value).Equals("\xE735") ^ (parameter as string ?? string.Empty).Equals("Reverse");
+        (Visibility)value == Visibility.Visible ^ (parameter as string ?? string.Empty).Equals("Reverse");
+
   }
 }
