+++ StudySmarterFlashcards/Menus/SettingsPage.xaml
@@ -2,14 +2,14 @@
   x:Class="StudySmarterFlashcards.Menus.SettingsPage"
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:utils="using:StudySmarterFlashcards.Utils"
+  xmlns:converters="using:StudySmarterFlashcards.Converters"
   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:interact="using:Microsoft.Xaml.Interactivity" xmlns:interactcore="using:Microsoft.Xaml.Interactions.Core"
       DataContext="{Binding SettingsInstance, Source={StaticResource Locator}}"
   mc:Ignorable="d"
   Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
   <Page.Resources>
-    <utils:BooleanToVisibilityConverter x:Key="boolToVisibilityConverter"/>
+    <converters:BooleanToVisibilityConverter x:Key="boolToVisibilityConverter"/>
   </Page.Resources>
   
   <Grid>
