+++ StudySmarterFlashcards/Sets/SetPage.xaml
@@ -6,17 +6,17 @@
     xmlns:interactcore="using:Microsoft.Xaml.Interactions.Core"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
-    xmlns:utils="using:StudySmarterFlashcards.Utils"
+    xmlns:converters="using:StudySmarterFlashcards.Converters"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     DataContext="{Binding SetInstance, Source={StaticResource Locator}}"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
   <Page.Resources>
-    <utils:BooleanToBackgroundConverter x:Key="boolToBackgroundConverter"/>
-    <utils:BooleanToVisibilityConverter x:Key="boolToVisibilityConverter"/>
-    <utils:StringFormatConverter x:Key="DateStringWithTimeConverter" StringFormat="\{0\}: \{1:dd MMMM yyyy h:mm tt\}" Title="Last Reviewed" ConvertToLocalTime="True"/>
-    <utils:StringFormatConverter x:Key="DateStringNoTimeConverter" StringFormat="\{0\}: \{1:dd MMMM yyyy\}" Title="Created" ConvertToLocalTime="False"/>
-    <utils:StringFormatConverter x:Key="NumTimesReviewdConverter" StringFormat="\{0\}: \{1\}" Title="Number of Times Reviewed"/>
+    <converters:BooleanToBackgroundConverter x:Key="boolToBackgroundConverter"/>
+    <converters:BooleanToVisibilityConverter x:Key="boolToVisibilityConverter"/>
+    <converters:StringFormatConverter x:Key="DateStringWithTimeConverter" StringFormat="\{0\}: \{1:dd MMMM yyyy h:mm tt\}" Title="Last Reviewed" ConvertToLocalTime="True"/>
+    <converters:StringFormatConverter x:Key="DateStringNoTimeConverter" StringFormat="\{0\}: \{1:dd MMMM yyyy\}" Title="Created" ConvertToLocalTime="False"/>
+    <converters:StringFormatConverter x:Key="NumTimesReviewdConverter" StringFormat="\{0\}: \{1\}" Title="Number of Times Reviewed"/>
   </Page.Resources>
   <Grid>
     <Grid.RowDefinitions>
