+++ StudySmarterFlashcards/StudySmarterFlashcards.csproj
@@ -157,17 +157,17 @@
     </Compile>
     <Compile Include="Study\BasicStudyViewModel.cs" />
     <Compile Include="Utils\BaseViewModel.cs" />
-    <Compile Include="Utils\BooleanToBackgroundConverter.cs" />
-    <Compile Include="Utils\BooleanToStarFillConverter.cs" />
-    <Compile Include="Utils\BooleanToVisibilityConverter.cs" />
-    <Compile Include="Utils\EqualityToVisibilityConverter.cs" />
+    <Compile Include="Converters\BooleanToBackgroundConverter.cs" />
+    <Compile Include="Converters\BooleanToStarFillConverter.cs" />
+    <Compile Include="Converters\BooleanToVisibilityConverter.cs" />
+    <Compile Include="Converters\EqualityToVisibilityConverter.cs" />
     <Compile Include="Utils\IErrorHandler.cs" />
     <Compile Include="Utils\LocalDataHandler.cs" />
     <Compile Include="Dialogs\StudyInstructionsDialog.xaml.cs">
       <DependentUpon>StudyInstructionsDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="Utils\NotifyTaskCompletion.cs" />
-    <Compile Include="Utils\StringFormatConverter.cs" />
+    <Compile Include="Converters\StringFormatConverter.cs" />
     <Compile Include="Utils\TaskUtils.cs" />
     <Compile Include="Utils\ViewModelLocator.cs" />
   </ItemGroup>
